Rationalise the use of flot.
The previous code had lots of custom functions, hardcoded javascript strings
and various layering violations.  This is fixed so objects produce their own
data, this is passed via JSON on request to the client who plots it on
demand in a (mostly) general way.

Several bits of code relating to graphs which never worked have been
removed.  If they are deemed useful we can add them again.